+++ src/os2datascanner/projects/admin/import_services/models/google_workspace_configuration.py
@@ -3,6 +3,7 @@ from django.db import models
 from .import_service import ImportService
 from .exported_mixin import Exported
 from django.utils.translation import gettext_lazy as _
+from os2datascanner.projects.admin.core.models.background_job import JobState
 
 
 class GoogleWorkspaceConfig(Exported, ImportService):
@@ -28,3 +29,24 @@ class GoogleWorkspaceConfig(Exported, ImportService):
     class Meta:
         verbose_name = _("Google Workspace configuration")
         verbose_name_plural = _("Google Workspace configurations")
+
+    def start_import(self):
+        from . import GoogleWorkspaceImportJob
+        from os2datascanner.projects.admin.import_services.utils import _start_import_job
+
+        _start_import_job(
+            importjob_model=GoogleWorkspaceImportJob,
+            lookup_filter={"grant": self.grant},
+            job_kwargs={
+                "grant": self.grant,
+                "organization": self.organization,
+                "delegated_admin_email": self.delegated_admin_email,
+            },
+            allowed_states=(
+                JobState.FINISHED,
+                JobState.FAILED,
+                JobState.CANCELLED,
+                JobState.FINISHED_WITH_WARNINGS,
+            ),
+            log_name=f"GoogleWorkspaceConfig {self.pk}",
+        )